+++ mongol-library/src/main/java/net/studymongolian/mongollibrary/MongolTextView.java
@@ -256,7 +256,7 @@ public class MongolTextView extends View  implements ViewTreeObserver.OnPreDrawL
     /**
      * @param size in SP units
      */
-    public void setTextSize(int size) {
+    public void setTextSize(float size) {
         mTextSizePx = TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
                 size, getResources().getDisplayMetrics());
         mTextPaint.setTextSize(mTextSizePx);
